Move DTO to Post logic from the mongo service into a Post factory method
